requirements.txt
matplotlib~=3.10.7
numpy~=2.1.2
jieba3k~=0.35.1
pandas~=2.3.3
wordcloud~=1.9.4
requests~=2.32.5
lxml~=6.0.2
transformers~=4.57.1